- updated the makefiles and code to allow compliation with MinGW once again (some of patch #875485 by perditionc) - fixed errors and warnings given by gcc
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@3513 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/ShConstants.h
@@ -38,7 +38,7 @@ class ConstantsStringList : public SortedStringListND<struct constantstring>
       return temp;
     }
 
-    int get(char *name, size_t n_chars = -1)
+    int get(char *name, int n_chars = -1)
     {
       int v=SortedStringListND<struct constantstring>::find(name, n_chars);
       if (v==-1) return -1;
